Change menu sliders to spin boxes
**Notes**
- All menu sliders are spin boxes

**Testing**
- `./style_check.sh`

**Issues**
https://github.com/dtsullivan/optgui/issues/41#issue-550585110

+++ Optimization_Interface/include/graphics/view.h
@@ -36,7 +36,7 @@ class View : public QGraphicsView {
  private slots:
     void openMenu();
     void closeMenu();
-    void setZoom(int value);
+    void setZoom(double value);
     void setState(STATE button_type);
     void loadFile();
     void saveFile();
